backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
drf-spectacular>=0.27
python-dotenv>=1.0
Pillow>=10.3
psycopg[binary]>=3.1
